requirements.txt
numpy==1.26.4
protobuf==4.25.3
scikit-learn==1.5.1
torch==2.4.1
tqdm==4.66.5
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-probability==0.19.0
